+++ app/renderer/types.js
@@ -48,6 +48,8 @@ const actionTypes = {
   stakeAndRestartNode: triggerActionTypes('stake_and_restart_node'),
   stakingStopStream: apiActionTypes('staking_stop_stream'),
   stakingRestartNode: apiActionTypes('staking_restart_network'),
+  stakingUnStake: apiActionTypes('staking_un_stake'),
+  stakingSavePreferences: apiActionTypes('staking_save_preferences'),
 
   /* system apis */
   nodeJsonRpcSystem: apiActionTypes('node_jsonrpc_system'),